requirements.txt
# ============================================================
# BiasGuard Dependencies
# ============================================================

# ── Core LLM & Agent Framework ─────────────────────────────
langchain==0.2.16
langchain-core==0.2.40
langchain-community==0.2.16
langchain-anthropic==0.1.23
langchain-openai==0.1.25
langchain-groq==0.1.10
langgraph==0.2.34

# ── Vector Database ────────────────────────────────────────
chromadb==0.5.15
pinecone-client==4.1.2

# ── Embeddings ─────────────────────────────────────────────
sentence-transformers==3.1.1
tiktoken==0.7.0

# ── API ────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
httpx==0.27.2

# ── Data Validation ────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2

# ── Document Processing ────────────────────────────────────
pypdf==5.0.1
python-docx==1.1.2
unstructured==0.15.14

# ── Report Generation ──────────────────────────────────────
markdown==3.7
weasyprint==62.3
jinja2==3.1.4

# ── Observability ──────────────────────────────────────────
langsmith==0.1.131
arize-phoenix==4.29.0
arize-phoenix-otel==0.5.1
opentelemetry-sdk==1.27.0
prometheus-client==0.21.0

# ── Utilities ──────────────────────────────────────────────
python-dotenv==1.0.1
structlog==24.4.0
tenacity==9.0.0
rich==13.8.1
typer==0.12.5

# ── Frontend ───────────────────────────────────────────────
streamlit==1.38.0
plotly==5.24.1
pandas==2.2.3
